chore(checkout): BACK-714 update switch to toggle display
