+++ packages/core/src/scss/components/bigcommerce/switch/_component.scss
@@ -1,5 +1,5 @@
 // =============================================================================
-// SWITCH COMPONENT
+// TOGGLE COMPONENT
 // =============================================================================
 
-@import "switch";
+@import "toggle";
