+++ packages/ui/src/switch/index.ts
@@ -1,1 +1,1 @@
-export { Switch } from './switch';
+export { Switch } from './toggle';
